services/prediction-engine/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
structlog>=23.1.0
google-cloud-firestore>=2.13.0
pydantic>=2.5.0
